Scalable marshal and unmarshal benchmarks.
Dear reader, I am so excited about this.

Also: added additional sanity checks to the end of benchmarks, after
a call to `b.StopTimer`.  Always important to make sure your benchmark
is actually doing what you think it is, and not being super fast
because it turns out to be a hyperspeed application-layer devnull!

(In the case of the unmarshal tests, asserting sanity by remarshalling
the whole dang thing *again* might look a little funny... but it's
certainly a terse way to get full coverage of the data!)