Port scaled marshal tests to new node interfaces.
Marshal is 25% faster.  Not bad.  (There's also a perf bug in the codec
code which I'm about to fix which will change the speedup to *double*.)

Unmarshal is a bit faster, but not as much improved as marshal.
This is unsurprisingly because we chose not to amortize allocations
as much as we could in map and list creation because the size penalty
seemed too steep to make it a worthy trade.  (This gives us *much*
better numbers to fuel further consideration of that trade, though.)