Expose styles with nicer syntax.
I'm not sure if there's a name for this way of grouping methods,
but the end result is dotted.Access.Patterns, and it's kinda nice.

This was extensively discussed in the PR:

https://github.com/ipld/go-ipld-prime/pull/49#issuecomment-596465234
https://github.com/ipld/go-ipld-prime/pull/49#issuecomment-597549108
https://github.com/ipld/go-ipld-prime/pull/49#issuecomment-604412736

The "inlinability_test.go" file can be compiled with special flags
(described in the comment at the top of the file) to see the
outcome in assembly.  Result?  Yep, things are still inlinable;
this change is performance neutral.